feat: option for optional deleteMessage and deleteOnUpdate

+++ src/events/onMessageUpdate.js
@@ -3,7 +3,7 @@
 const { deconstructMention } = require('../enhancedMention');
 const { triggerWH, setInMap, MESSAGE_LIMIT } = require('../utils');
 
-exports.onMessageUpdate = async(botClient, network, channelsCache, msg, oldMsg) => {
+exports.onMessageUpdate = async(botClient, network, channelsCache, deleteOnUpdate, msg, oldMsg) => {
     // !oldMsg = message not cached = don't log the update
     if (!msg.author || msg.author.bot || !msg.channel.guild || !oldMsg) {
         return;
@@ -39,10 +39,11 @@ exports.onMessageUpdate = async(botClient, network, channelsCache, msg, oldMsg) 
         }
         messages.push(await triggerWH(botClient, network, channelConfig, cur, msg.author, fullMsg) );
     }
-    const toDelete = channelsCache[msg.channel.id].get(msg.id);
-    for (const m of toDelete) {
-        m.delete().catch();
+    if (deleteOnUpdate) {
+        const toDelete = channelsCache[msg.channel.id].get(msg.id);
+        for (const m of toDelete) {
+            m.delete().catch();
+        }
     }
-
     setInMap(channelsCache[msg.channel.id], msg.id, messages);
 };
